refactor: move dark-mode-state hook into other directory
