+++ frontend/src/components/editor-page/editor-pane/linter/linter.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 import { allAppExtensions } from '../../../../extensions/all-app-extensions'
-import { useDarkModeState } from '../../../../hooks/common/use-dark-mode-state'
+import { useDarkModeState } from '../../../../hooks/dark-mode/use-dark-mode-state'
 import { FrontmatterLinter } from './frontmatter-linter'
 import type { Diagnostic } from '@codemirror/lint'
 import { linter } from '@codemirror/lint'
