+++ frontend/src/extensions/external-lib-app-extensions/flowchart/flowchart.tsx
@@ -7,7 +7,7 @@ import fontStyles from '../../../../global-styles/variables.module.scss'
 import { AsyncLoadingBoundary } from '../../../components/common/async-loading-boundary/async-loading-boundary'
 import { ShowIf } from '../../../components/common/show-if/show-if'
 import type { CodeProps } from '../../../components/markdown-renderer/replace-components/code-block-component-replacer'
-import { useDarkModeState } from '../../../hooks/common/use-dark-mode-state'
+import { useDarkModeState } from '../../../hooks/dark-mode/use-dark-mode-state'
 import { Logger } from '../../../utils/logger'
 import { testId } from '../../../utils/test-id'
 import React, { useEffect, useRef, useState } from 'react'
